+++ CHANGELOG.md
@@ -7,3 +7,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## Added
 - JIT evaluation loop.
+- VM options control.